+++ library/Database/Exception.php
@@ -1,0 +1,24 @@
+<?php
+
+/**
+ * This file is part of slick/database package
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Slick\Database;
+
+/**
+ * Generic exception interface for slick/database package
+ *
+ * All slick/common exceptions implement this interface to make it possible
+ * to identify them in your code.
+ *
+ * @package Slick\Database
+ * @author  Filipe Silva <filipe.silva@sata.pt>
+ */
+interface Exception
+{
+
+}