+++ library/Database/Exception/ServiceException.php
@@ -1,0 +1,24 @@
+<?php
+
+/**
+ * This file is part of slick/database package
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Slick\Database\Exception;
+
+use RuntimeException;
+use Slick\Database\Exception;
+
+/**
+ * Database service related and runtime errors
+ *
+ * @package Slick\Database\Exception
+ * @author  Filipe Silva <filipe.silva@sata.pt>
+ */
+class ServiceException extends RuntimeException implements Exception
+{
+
+}