+++ tests/Database/Fixtures/CustomIterator.php
@@ -1,0 +1,23 @@
+<?php
+
+/**
+ * This file is part of slick/database package
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Slick\Tests\Database\Fixtures;
+
+use ArrayIterator;
+
+/**
+ * Custom Iterator for tests
+ *
+ * @package Slick\Tests\Database\Fixtures
+ * @author  Filipe Silva <filipe.silva@sata.pt>
+ */
+class CustomIterator extends ArrayIterator
+{
+
+}